Add [CEReactions] annotations
This is part of the effort in https://github.com/w3c/webcomponents/issues/186 to annotate all parts of the platform that cause DOM mutations and thus need to set up state for custom element reactions. [CEReactions] is defined in https://html.spec.whatwg.org/multipage/scripting.html#cereactions.

Also normalizes whitespace to spaces instead of a mix of spaces and tabs.