Only generate an attribute prefix if required
If implemented as specified, the prefix set for an attribute in the DOM
would never be preserved during serialization. This change only
generates the new prefix if the prefix is either missing (which is not
allowed for namespaced attributes), or if it conflicts with another
local namespace definition of the prefix for a different namespace.

Fixes #29